refactor: use better name for user storage directory
